requirements.txt
Flask==3.1.2
geoip2==5.1.0
python-dotenv==1.1.1
Requests==2.32.5
user_agents==2.2.0
